+++ Event/ControllerArgumentsEvent.php
@@ -53,7 +53,7 @@ final class ControllerArgumentsEvent extends KernelEvent
     }
 
     /**
-     * @param array<class-string, list<object>>|null $attributes
+     * @param list<object>|null $attributes
      */
     public function setController(callable $controller, ?array $attributes = null): void
     {
@@ -99,9 +99,9 @@ final class ControllerArgumentsEvent extends KernelEvent
     /**
      * @template T of object
      *
-     * @param class-string<T>|null $className
+     * @param class-string<T>|'*'|null $className
      *
-     * @return ($className is null ? array<class-string, list<object>> : list<T>)
+     * @return ($className is null ? array<class-string, list<object>> : ($className is '*' ? list<object> : list<T>))
      */
     public function getAttributes(?string $className = null): array
     {